requirements.txt
boto3~=1.34.0
click~=8.1.0
PyYAML~=6.0
Flask~=3.0
